assignment_1/task_1/grupo_5/gabriel_rodrigues/requirements.txt
mysql-connector-python
python-dotenv